Make v2 tests less aggressive (#578)
Currently it is DDoSing the arangod in single and active failover mode such that arangod can't handle the creation of  all DBs/Collections (it tries to update user permissions but fails after 10 attempts). This happens on nodes with high number of cores, e.g. 36 because default value of parallelism is GOMAXPROCS(=Num of cores)

It does not happen with v1 tests because we don't try to execute so many tests in parallel there.